Extract card-document shape predicates to decorator-free subpath
PR-review followup: use the existing `isSingleCardDocument` helper
instead of hand-rolling the shape guard in `prepareExampleInstance`.

The runtime-common helper lives in `document-types.ts`, but importing
that file transitively loads `resource-types.ts` → `code-ref.ts` →
`loader.ts` → `realm.ts` → `realm-index-query-engine.ts`, which uses
`@Memoize()` decorators. Playwright's TS loader compiles decorators via
the TC39 protocol and rejects `typescript-memoize`, so the
software-factory Playwright harness can't pull any module reachable
through that chain. (This is the same constraint that already produced
`supported-mime-type` / `matrix-constants` subpaths in runtime-common.)

Applies the decorator-free subpath pattern for this case:

- New `packages/runtime-common/card-document-shape.ts` holds the pure
  shape predicates — `isCodeRef`, `isResolvedCodeRef`, `isRelationship`,
  `isCardFields`, `isMeta`, `isCardResource`, `isFileMetaResource`,
  `isSingleCardDocument`, `isCardCollectionDocument` — with only `import
  type` dependencies on the heavier neighbors. No runtime import reaches
  `loader.ts` or the decorator chain.
- `code-ref.ts` / `resource-types.ts` / `document-types.ts` now
  re-export these predicates from `card-document-shape.ts` (backward
  compat — every existing import path keeps working).
- `packages/software-factory/src/instantiate-execution.ts` imports
  `isSingleCardDocument` and `isResolvedCodeRef` via
  `@cardstack/runtime-common/card-document-shape` and uses them directly
  to guard `prepareExampleInstance`. Drops the hand-rolled shape-check
  helper.
