+++ packages/runtime-common/resource-types.ts
@@ -1,5 +1,5 @@
 import type { RealmInfo } from './realm';
-import { type CodeRef, isCodeRef, moduleFrom } from './code-ref';
+import { type CodeRef, moduleFrom } from './code-ref';
 import { resolveCardReference } from './card-reference-resolver';
 import type { Query } from './query';
 
@@ -145,203 +145,17 @@ export function isModuleResource(resource: any): resource is ModuleResource {
   return resource.type === 'source';
 }
 
-//validation - cards
-export function isCardResource(resource: any): resource is CardResource {
-  if (typeof resource !== 'object' || resource == null) {
-    return false;
-  }
-  if ('id' in resource && typeof resource.id !== 'string') {
-    return false;
-  }
-  if ('lid' in resource && typeof resource.lid !== 'string') {
-    return false;
-  }
-  if ('type' in resource && resource.type !== CardResourceType) {
-    return false;
-  }
-  if ('attributes' in resource && typeof resource.attributes !== 'object') {
-    return false;
-  }
-  if ('relationships' in resource) {
-    let { relationships } = resource;
-    if (typeof relationships !== 'object' || relationships == null) {
-      return false;
-    }
-    for (let [fieldName, relationship] of Object.entries(relationships)) {
-      if (typeof fieldName !== 'string') {
-        return false;
-      }
-      if (Array.isArray(relationship)) {
-        if (relationship.some((entry) => !isRelationship(entry))) {
-          return false;
-        }
-      } else if (!isRelationship(relationship)) {
-        return false;
-      }
-    }
-  }
-  if (!('meta' in resource) || typeof resource.meta !== 'object') {
-    return false;
-  }
-  let { meta } = resource;
-
-  if ('fields' in meta) {
-    if (!isCardFields(meta.fields)) {
-      return false;
-    }
-  }
-
-  if (!('adoptsFrom' in meta) || typeof meta.adoptsFrom !== 'object') {
-    return false;
-  }
-  let { adoptsFrom } = meta;
-  return isCodeRef(adoptsFrom);
-}
-
-export function isFileMetaResource(
-  resource: any,
-): resource is FileMetaResource {
-  if (typeof resource !== 'object' || resource == null) {
-    return false;
-  }
-  if ('id' in resource && typeof resource.id !== 'string') {
-    return false;
-  }
-  if (!('type' in resource) || resource.type !== FileMetaResourceType) {
-    return false;
-  }
-  if ('attributes' in resource && typeof resource.attributes !== 'object') {
-    return false;
-  }
-  if ('relationships' in resource) {
-    let { relationships } = resource;
-    if (typeof relationships !== 'object' || relationships == null) {
-      return false;
-    }
-    for (let [fieldName, relationship] of Object.entries(relationships)) {
-      if (typeof fieldName !== 'string') {
-        return false;
-      }
-      if (Array.isArray(relationship)) {
-        if (relationship.some((entry) => !isRelationship(entry))) {
-          return false;
-        }
-      } else if (!isRelationship(relationship)) {
-        return false;
-      }
-    }
-  }
-  if (!('meta' in resource) || typeof resource.meta !== 'object') {
-    return false;
-  }
-  let { meta } = resource;
-
-  if ('fields' in meta) {
-    if (!isCardFields(meta.fields)) {
-      return false;
-    }
-  }
-
-  if (!('adoptsFrom' in meta) && typeof meta.adoptsFrom !== 'object') {
-    return false;
-  }
-  let { adoptsFrom } = meta;
-  return isCodeRef(adoptsFrom);
-}
-
-export function isCardFields(fields: any): fields is CardFields {
-  if (typeof fields !== 'object') {
-    return false;
-  }
-  for (let [fieldName, fieldItem] of Object.entries(
-    fields as { [fieldName: string | symbol]: any },
-  )) {
-    if (typeof fieldName !== 'string') {
-      return false;
-    }
-    if (Array.isArray(fieldItem)) {
-      if (fieldItem.some((f) => !isMeta(f, true))) {
-        return false;
-      }
-    } else if (!isMeta(fieldItem, true)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-export function isMeta(meta: any, allowPartial: true): meta is Partial<Meta>;
-export function isMeta(meta: any): meta is Meta;
-export function isMeta(meta: any, allowPartial = false) {
-  if (typeof meta !== 'object' || meta == null) {
-    return false;
-  }
-  if ('adoptsFrom' in meta) {
-    let { adoptsFrom } = meta;
-    if (!isCodeRef(adoptsFrom)) {
-      return false;
-    }
-  } else {
-    if (!allowPartial) {
-      return false;
-    }
-  }
-  if ('fields' in meta) {
-    if (!isCardFields(meta.fields)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-export function isRelationship(
-  relationship: any,
-): relationship is Relationship {
-  if (typeof relationship !== 'object' || relationship == null) {
-    return false;
-  }
-  if ('meta' in relationship && typeof relationship.meta !== 'object') {
-    return false;
-  }
-  if ('links' in relationship) {
-    let { links } = relationship;
-    if (typeof links !== 'object' || links == null) {
-      return false;
-    }
-    if (!('self' in links)) {
-      return false;
-    }
-    let { self } = links;
-    if (typeof self !== 'string' && self !== null) {
-      return false;
-    }
-    if ('related' in links) {
-      if (typeof links.related !== 'string' && links.related !== null) {
-        return false;
-      }
-    }
-  } else if ('data' in relationship) {
-    let { data } = relationship;
-    if (typeof data !== 'object') {
-      return false;
-    }
-    if (data !== null && 'type' in data && 'id' in data) {
-      let { type, id } = data;
-      if (typeof type !== 'string' || typeof id !== 'string') {
-        return false;
-      }
-    }
-    if (data !== null && 'type' in data && 'lid' in data) {
-      let { type, lid } = data;
-      if (typeof type !== 'string' || typeof lid !== 'string') {
-        return false;
-      }
-    }
-  } else {
-    return false;
-  }
-  return true;
-}
+// Pure shape predicates live in `card-document-shape.ts` so callers that
+// only need to recognize a JSON:API resource don't pull the transitive
+// runtime chain rooted in this file (`card-reference-resolver.ts` →
+// `loader.ts` → `realm.ts` → ...). Re-exported here for backward compat.
+export {
+  isCardResource,
+  isFileMetaResource,
+  isCardFields,
+  isMeta,
+  isRelationship,
+} from './card-document-shape';
 
 export function extractRelationshipIds(
   relationship: Relationship,
